requirements.txt
flask==2.3.3
pillow==10.0.0
pydantic==2.3.0
python-dotenv==1.0.0
uuid==1.30
pytest==7.4.0
Flask-WTF==1.1.1
email-validator==2.0.0
pytesseract==0.3.10
opencv-python==4.8.0.74
numpy==1.24.3
requests==2.31.0
google-cloud-vision==3.4.3